#3311: move method to util directory; add tests
