+++ src/hooks/useDeployments.ts
@@ -30,13 +30,15 @@ import { refreshRegistries } from "@/hooks/useRefresh";
 import { Dispatch } from "redux";
 import { mergePermissions } from "@/utils/permissions";
 import { Permissions } from "webextension-polyfill";
-import { IExtension, UUID, RegistryId } from "@/core";
+import { IExtension, RegistryId, UUID } from "@/core";
 import { maybeGetLinkedApiClient } from "@/services/apiClient";
-import { satisfies } from "semver";
-import { compact, gte } from "lodash";
 import chromeP from "webext-polyfill-kinda";
 import extensionsSlice from "@/store/extensionsSlice";
 import useFlags from "@/hooks/useFlags";
+import {
+  checkExtensionUpdateRequired,
+  makeUpdatedFilter,
+} from "@/utils/deployment";
 
 const { actions } = extensionsSlice;
 
@@ -75,61 +77,6 @@ async function fetchDeployments(
 
   return deployments;
 }
-
-/**
- * Returns a predicate that returns `true` if a deployment requires an update.
- *
- * Restricted users:
- * - The remote deployment's extension updated_at timestamp is after the installed version. The updated_at timestamp
- *   can change even if the blueprint version doesn't change because of 1) the deployment was paused, or 2) the
- *   admin changed a binding on the deployment
- *
- * Unrestricted users:
- * - Same as above, but ignore deployments where the user has a newer version of the blueprint installed because that
- *   means they are doing local deployment on the blueprint.
- *
- * @param installed the user's currently installed extensions (including for paused deployments)
- * @param restricted `true` if the user is a restricted organization user (i.e., as opposed to a developer)
- */
-export const makeUpdatedFilter =
-  (installed: IExtension[], { restricted }: { restricted: boolean }) =>
-  (deployment: Deployment) => {
-    const deploymentMatch = installed.find(
-      (extension) => extension._deployment?.id === deployment.id
-    );
-
-    if (restricted) {
-      return (
-        !deploymentMatch ||
-        new Date(deploymentMatch._deployment.timestamp) <
-          new Date(deployment.updated_at)
-      );
-    }
-
-    // Local copies an unrestricted user (i.e., a developer role) is working on
-    const blueprintMatch = installed.find(
-      (extension) =>
-        extension._deployment == null &&
-        extension._recipe.id === deployment.package.package_id
-    );
-
-    if (!deploymentMatch && !blueprintMatch) {
-      return true;
-    }
-
-    if (
-      blueprintMatch &&
-      gte(blueprintMatch._recipe.version, deployment.package.version)
-    ) {
-      // The unrestricted user already has the blueprint (or a newer version of the blueprint), so don't prompt
-      return false;
-    }
-
-    return (
-      new Date(deploymentMatch._deployment.timestamp) <
-      new Date(deployment.updated_at)
-    );
-  };
 
 function activateDeployments(
   dispatch: Dispatch,
@@ -201,30 +148,6 @@ export type DeploymentState = {
   error: unknown | undefined;
 };
 
-/**
- * Returns `true` if the user needs to update their browser extension to install the deployments.
- *
- * For now assumes only minimum version requirements. (I.e., this method also returns true if there's a maximum version
- * violation).
- */
-function checkExtensionUpdateRequired(deployments: Deployment[]): boolean {
-  // Check that the user's extension van run the deployment
-  const { version: extensionVersion } = browser.runtime.getManifest();
-  const versionRanges = compact(
-    deployments.map((x) => x.package.config.metadata.extensionVersion)
-  );
-
-  console.debug("Checking deployment version requirements", {
-    versionRanges,
-    extensionVersion,
-  });
-
-  // For now assume the only version restrictions will be that the version must be greater than a certain number
-  return versionRanges.some(
-    (versionRange) => !satisfies(extensionVersion, versionRange)
-  );
-}
-
 function useDeployments(): DeploymentState {
   const dispatch = useDispatch();
   const installedExtensions = useSelector(selectExtensions);
@@ -244,7 +167,7 @@ function useDeployments(): DeploymentState {
       updatedDeployments,
       checkExtensionUpdateRequired(updatedDeployments),
     ];
-  }, [installedExtensions, deployments]);
+  }, [restrict, installedExtensions, deployments]);
 
   const handleUpdate = useCallback(async () => {
     if (deployments == null) {
